bridges-manager removed and function added into get-tor-bridges. and code fixed.

+++ get-tor-bridges.sh
@@ -1,58 +1,43 @@
 #!/usr/bin/env bash
-
-
-HTML_FILE="img.html"
-IMAGE_FILE="captcha.jpg"
-BridgeFile="bridges.txt"
-IpSite="icanhazip.com"
-
 # Add config file 
 source tbcli-config 2>/dev/null
+
 if [[ $? -ne 0 ]] ; then 
-    echo -e "\e[31mfile config Not exist" 
+    echo -e "\e[1;31mfile tbcli_config Not exist" 
     exit 1 
 fi 
 
 # requirements for run this script
 if [[ ! $(which feh) || ! $(which tor) || ! $(which proxychains4) || ! $(which obfs4proxy)  ]]; then
-    echo -e "${light_red}Check that programs are installed ? => ( proxychains4 , feh ,tor , obfs4proxy )\e[m "
+    echo -e "${light_red}Check that programs are installed ? => ( proxychains4 , feh ,tor , obfs4proxy )${nc}"
     exit 1
 fi
 
 # usage
 function usage(){
-    echo -e "${light_yellow}A Simple Script for get tor bridge from${light_magenta} https://bridges.torproject.org/bridges\e[m"
+    echo -e "${light_yellow}A Simple Script for get tor bridge from${light_magenta} ${url_bridges}${nc}"
     echo -e "${light_magenta}USAGE :"
-    echo -e "${light_green}\t-a | --add-bridges\t\t${cyan}add bridges to $tor_config_file"
-    echo -e "${light_green}\t-d | --disable-broken-bridge\t${cyan}Disable broken Bridges in this network connection"
-    echo -e "${light_green}\t-e | --enable-all-bridge\t${cyan}Enable all disabled Bridges"
-    echo -e "${light_green}\t-r | --reset-tor\t\t${cyan}restart tor service"
-    echo -e "${light_green}\t-u | --uninstall\t\t${cyan}uninstall Script"
-    echo -e "${light_green}\t-h | --help\t\t\t${cyan}show this help"
+    echo -e "${light_green}\t-a | -A | --add-bridges\t\t\t${cyan}add bridges into /etc/tor/torrc and print bridges"
+    echo -e "${light_green}\t-p | -P | --print-only-bridges\t\t${cyan}just print bridges" 
+    echo -e "${light_green}\t-d | -D | --disable-broken-bridges\t${cyan}Disable broken Bridges in this network connection"
+    echo -e "${light_green}\t-e | -E | --enable-all-bridges\t\t${cyan}Enable all disabled Bridges"
+    echo -e "${light_green}\t-r | -R | --reset-tor\t\t\t${cyan}restart tor service"
+    echo -e "${light_green}\t-u | -U | --uninstall\t\t\t${cyan}uninstall Script"
+    echo -e "${light_green}\t-h | -H | --help\t\t\t${cyan}show this help"
     echo -e "${nc}"
 }
 
-# delete files
-function ClearFiles() {
-    for file in "${HTML_FILE}" "${IMAGE_FILE}" "${BridgeFile}"; do
+# delete tmp files
+function ClearTmpFiles() {
+    for file in "${HTML_FILE}" "${IMAGE_CAPTCHA_FILE}" "${BridgeFile}"; do
         if [[ -e "$file" ]]; then
             rm $file
         fi
     done
 }
 
-# get tor bridges
+# get tor bridges from 
 function get_tor_bridges(){
-
-    # add bridges into file $tor_config_file (default=False)
-    add_bridges="$1"
-
-    # args for manager bridges 
-    args_bridges_manager="$2"
-
-    # delete tmp files
-    ClearFiles 
-
     # get Captcha
     proxychains4 -q curl -s "$url_bridges" -o "$HTML_FILE"
 
@@ -60,59 +45,64 @@ function get_tor_bridges(){
     [[ -z $(cat "$HTML_FILE" 2>/dev/null) ]] && echo "Error TOR" && exit 0
 
     # cut base64 Image challenge and convert to image
-    base64 -i -d <<< $(cat "$HTML_FILE" |egrep -o "\/9j\/[^\"]*") > $IMAGE_FILE
+    base64 -i -d <<< $(cat "$HTML_FILE" |egrep -o "\/9j\/[^\"]*") > $IMAGE_CAPTCHA_FILE
 
-    # view image
-    feh "$IMAGE_FILE" &
+    # show image captcha security code
+    feh "$IMAGE_CAPTCHA_FILE" &
     FEH_PID=$!
 
-    # Captcha challenge field ( code of captcha )
-    Cap_Challenge=$(cat "$HTML_FILE" |grep value |head -n 1 |cut -d\" -f 2)
+    # Captcha challenge field ( captcha serial )
+    Cap_Serial=$(cat "$HTML_FILE" |grep value |head -n 1 |cut -d\" -f 2)
 
-    # Enter code captcha
-    while [[ -z $Cap_Response ]]; do
-        # show ip
+    # captcha security code
+    while [[ -z $captcha_security_code ]]; do
+        # show your ip
         echo -ne "${light_magenta}[ $(proxychains4 -q curl -s "$IpSite") ] ${light_blue}"
-
-
+        
         # get code from user
-        read -p "Enter code (Enter 'r' For Reset Captcha): " Cap_Response
+        read -p "Enter code (Enter 'r' For Reset Captcha): " captcha_security_code
 
-        # reset captcha
-        [[ $Cap_Response == "r" ]] &&
-        {
-            # kill feh job
+        # press 'r' for reset captcha
+        if [[ $captcha_security_code == "r" ]]; then 
+            # close image captcha security code 
             kill $FEH_PID
 
-            # clear data in Cap_Response for while condition true
-            Cap_Response=""
+            # delete tmp files
+            ClearTmpFiles 
+
+            # unset captcha_security_code
+            unset captcha_security_code
 
             # reset captcha
-            get_tor_bridges "$AddBridges" "$RemoveBrokenBridges" "$ReTor"
+            get_tor_bridges
 
             exit 0
-        }
+        fi
 
         # slove captcha and get Bridges
         proxychains4 -q curl -s "$url_bridges" \
-        --data "captcha_challenge_field=${Cap_Challenge}&captcha_response_field=${Cap_Response}&submit=submit" -o "$BridgeFile"
+        --data "captcha_challenge_field=${Cap_Serial}&captcha_response_field=${captcha_security_code}&submit=submit" -o "$BridgeFile"
 
         # cut bridges from html file(if code is incorrect bridges file is empty)
         RES=$(cat "$BridgeFile" |grep obfs4 |egrep -o "^[^<]*")
 
-        # if Cap_Response is correct. bridges save into $tor_config_file . incorrect show error
+        # if captcha_security_code is correct code save bridges into Variable BRIDGES
+        # if incorrect print error
         if [[ ! -z $(echo $RES|tr -d '\n') ]]; then
             BRIDGES=$(echo "$RES" |sed 's/^/Bridge /g')
         else
-            echo -e "${light_yellow}The code entered is incorrect! try again ..\e[m"
-            # continue . True while
-            Cap_Response=""
+            echo -e "${light_yellow}The code entered is incorrect! try again ..${nc}"
+            # for continue while and try again . unset captcha_security_code
+            unset captcha_security_code
         fi
     done
 
-    # kill feh job
+    # close image captcha security code 
     kill $FEH_PID
+}
 
+# add bridges into tor config file 
+function save_and_print_bridges(){
     # add Bridges Into torrc
     if [ -e "$tor_config_file" ]; then
 
@@ -127,22 +117,81 @@ function get_tor_bridges(){
             echo "ClientTransportPlugin obfs4 exec ${obfs4proxy}" >> $tor_config_file
         fi
 
-        if [[ "$add_bridges" == "True" ]]; then
-            sudo echo -e "\n${BRIDGES}" >> $tor_config_file
-            echo -e "${light_magenta}Bridges Added into ${magenta}${tor_config_file}\n"
-        fi
-
+        # add bridges
+        echo -e "\n${BRIDGES}" >> $tor_config_file
+        echo -e "${light_magenta}Bridges Added into ${magenta}${tor_config_file}: \n"
+        
         # show bridges
         echo -e "${cyan}${BRIDGES}"
-
-        if [[ ! -z "$args_bridges_manager" ]]; then
-            bridges-manager $args_bridges_manager
-        fi
 
     else
         echo -e "${red}Tor config file $tor_config_file doesn't exist"
         exit 0
     fi
+}
+
+# read 'unable connetion bridges' from status tor and comment broken bridges
+function disable_broken_bridges(){
+    # check tor file is exist
+    if [ -e "$tor_config_file" ]; then
+        echo -ne "${light_yellow}waiting for find broken bridges:\n\t"
+        # broken bridges
+        broken_bridges=$(systemctl status tor.service|grep "unable"|egrep -o "([0-9]{1,3}.){3}[0-9]{1,3}:[0-9]{2,8}")
+        # check bridges exist
+        [[ ! -z $(echo $broken_bridges|tr -d '\n') ]] &&
+            {
+                echo -e "${light_red}Broken Bridges:"
+                # comment broken bridges 
+                for bridge in ${broken_bridges[@]};do
+                    echo -e "\t${light_magenta}[${light_red}X${light_magenta}] ${cyan}${bridge}"
+                    sed -i "s/^Bridge.*${bridge}/#&/g" "$tor_config_file"
+                done
+                echo -e "${light_yellow}broken bridges successfully disable.${nc}"
+            } ||
+                echo -e "${magenta}All bridges are healthy${nc}"
+        echo -e "${cyan}Active Bridges : $(cat $tor_config_file |grep ^Bridge|wc -l)${nc}"
+    else
+        echo -e "${red}Tor config file $tor_config_file doesn't exist"
+        exit 0
+    fi
+}
+
+# enable all disabled bridge
+function enable_all_bridges(){
+    # check tor file is exist
+    if [ -e "$tor_config_file" ]; then
+        Disable_Bridges=$(cat $tor_config_file | egrep "#Bridge obfs4")
+        if [[ ! -z "$Disable_Bridges" ]]; then 
+            echo -e "${yellow}Disable Bridges : \n${light_red}" 
+            echo "$Disable_Bridges"| cud -d " " -f 3 
+            sed -i "s/^#Bridge obfs4/Bridge obfs4/g" "$tor_config_file" 
+            echo -e "${light_green}Enabled."
+        else
+            echo -e "${light_yellow}All bridges are enable"
+            echo -e "${cyan}Active Bridges : $(cat /etc/tor/torrc |grep ^Bridge|wc -l)${nc}"
+        fi
+    else
+        echo -e "${red}Tor config file $tor_config_file doesn't exist"
+        exit 0
+    fi
+    
+}
+
+# restart tor with display bar 
+function reset_tor(){
+    # restart tor
+    systemctl restart tor.service
+    echo -e "${magenta}wait for restart tor service .."
+    # my tor status bar
+    res=0
+    status=" "
+    sep="${reset_sep}"
+    while [[ "$res" -lt "100" ]];do
+        sleep 0.1
+        res=$(systemctl status tor.service |egrep -o "Bootstrapped[^%]*"|tail -n 1|cut -d' ' -f2)
+        [[ -z $(grep "$res" <<< "$status") ]] && status="$res$status" && echo -ne "${sep}${res}"
+    done
+    echo -e "\n"
 }
 
 # uninstall script
@@ -170,26 +219,25 @@ function UninstallTBCLI(){
     fi
     echo -e "${nc}"
 }
-# # reset terminal
-# tput reset
 
 
-add_bridges="False"
-args_bridges_manager=""
+# reset terminal
+#tput reset
+
+# without args print help
+if [ $# -lt 1 ] ;then 
+    usage && exit 0
+fi 
 
 while [ "$1" != "" ]; do
     case $1 in
-        -a | --add-bridges  ) add_bridges="True" ;;
-
-        -d | --disable-broken-bridge ) args_bridges_manager="$args_bridges_manager -d" ;;
-
-        -e | --enable-all-bridge     ) args_bridges_manager="$args_bridges_manager -e" ;;
-
-        -r | --reset-tor    ) args_bridges_manager="$args_bridges_manager -r" ;;  
-        
-        -u | --uninstall    ) UninstallTBCLI && exit 0 ;;
-
-        -h | --help | *) usage && exit 0;;
+        -a | -A | --add-bridges            ) add_bridges="True" ;;
+        -p | -P | --print-only-bridges     ) print_bridges="True" ;;
+        -d | -D | --disable-broken-bridges ) bridges_manager="$bridges_manager -d" ;;
+        -e | -E | --enable-all-bridges     ) bridges_manager="$bridges_manager -e" ;;
+        -r | -R | --reset-tor              ) bridges_manager="$bridges_manager -r" ;;  
+        -u | -U | --uninstall              ) UninstallTBCLI && exit 0 ;;
+        -h | -H | --help | *               ) usage && exit 0;;
     esac
     shift
 done
@@ -201,7 +249,36 @@ if [[ $UID -ne 0 ]]; then
 fi
 
 # delete tmp files
-trap ClearFiles EXIT
+trap ClearTmpFiles EXIT
 
-# get tor bridges
-get_tor_bridges "$add_bridges" "$args_bridges_manager"
+# get tor bridgesg
+if [[ ! -z $print_bridges || ! -z $add_bridges ]] ; then 
+    BRIDGES=""
+    get_tor_bridges
+fi
+
+# print only bridges ( not add )
+if [[ ! -z $print_bridges && -z $add_bridges ]] ; then 
+    echo -e "${cyan}${BRIDGES}"
+fi 
+
+# add bridges into tor config file and print bridges 
+if [[ ! -z $add_bridges ]] ; then 
+    save_and_print_bridges 
+fi 
+
+# enable all bridges
+if grep -q "\-e" <<< $bridges_manager  ;then  
+    enable_all_bridges
+fi
+
+# disable broken bridges
+if grep -q "\-d" <<< $bridges_manager ;then
+    disable_broken_bridges
+fi
+
+# reset tor service with systemctl
+if grep -q "\-r" <<< $bridges_manager ;then
+    reset_tor
+fi
+
